Now the OverrideConfigProvider takes a vararg of config providers instead of loaders.
This makes the binding and reloadstrategy be separated in each provider so you can have a different streategy for each one. The problem was that you only had one reload strategy and for example a classpath resource file won't change so there is no need to reload them!

Now the provider has a new method of "contains" to check if there is instead of checking null later on